requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
sentence-transformers==2.7.0
huggingface-hub==0.36.2
PyMuPDF==1.27.2
qdrant-client==1.9.1
pydantic==2.12.5
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.26.4
jinja2==3.1.2
serpapi==1.0.2
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.1
langchain-core>=0.3.0,<0.4
langchain-groq>=0.2.0
matplotlib>=3.7.0
